+++ src/main/java/com/sesiones/sesiones_backend/repository/GradoRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sesiones.sesiones_backend.entity.Grado;
 
-public interface GradoRepository extends JpaRepository<Grado, Long> {
+public interface GradoRepository extends JpaRepository<Grado, Integer> {
 
-    Optional<Grado> findByIdAndNivelId(Long id, Long nivelId);
+    Optional<Grado> findByIdAndNivelId(Integer id, Integer nivelId);
 }
+
