+++ src/main/java/com/sesiones/sesiones_backend/service/ObtenerDocumentoCurriculoService.java
@@ -18,10 +18,11 @@ public class ObtenerDocumentoCurriculoService {
     private final SessionResponseMapper sessionResponseMapper;
 
     @Transactional(readOnly = true)
-    public DocumentoCurriculoResponse execute(Long documentoCurriculoId) {
+    public DocumentoCurriculoResponse execute(Integer documentoCurriculoId) {
         return documentoCurriculoRepository.findDetailedById(documentoCurriculoId)
             .map(sessionResponseMapper::toDocumentoCurriculoResponse)
             .orElseThrow(() -> new ResourceNotFoundException("No existe el documento curricular con id " + documentoCurriculoId));
     }
 }
+
